got the player to change meshes when spawned but have to fix adding to network list
wont let a client write to the addplayerlist so it's only registering one player and overwriting the host info with the client info when spawned. have to figure out how to handle adding players to list better. two lists? the one with tyhe player info spawned at submit button and the other on connect/network spawn?